requirements.txt
# PodcastGenerator dependencies

# For summarization
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99

# For LLM script generation
openai>=1.0.0

# For fact-checking
sentence-transformers>=2.2.0

# For text-to-speech
gTTS>=2.4.0
boto3>=1.34.0  # For AWS Polly if used

# Audio processing
pydub>=0.25.1

# Utilities
numpy>=1.20.0
tqdm>=4.60.0
python-dotenv>=0.19.0  # For loading environment variables from .env files 